+++ client/my-sites/customer-home/main.tsx
@@ -4,8 +4,8 @@ import { useState } from 'react';
 import DocumentHead from 'calypso/components/data/document-head';
 import Main from 'calypso/components/main';
 import { useGetDomainsQuery } from 'calypso/data/domains/use-get-domains-query';
+import { useShouldShowLaunchpadFirst } from 'calypso/landing/stepper/hooks/use-should-show-launchpad-first';
 import PageViewTracker from 'calypso/lib/analytics/page-view-tracker';
-import { useShouldShowLaunchpadFirst } from 'calypso/state/selectors/should-show-launchpad-first';
 import CelebrateLaunchModal from './components/celebrate-launch-modal';
 import { FullScreenLaunchpad } from './components/full-screen-launchpad';
 import HomeContent from './components/home-content';
